Code Working 5/4
All code working including rmsFail, but need to go back now and comment everything, and then unit test everything, especially rmsFail...and need to determine what/if any classes/functions/files are obscure.  After that, I will move back on to introducing first order equations for reduction propositions...I like this approach because I need to remember everything before the fragility framework is me trying to emulate SBD as accurately as I can as it is done today but with an automated process.  My PhD's focus is to make the process better and more conscientious of a reduction with a fragility check before committing to reductions.

The fragility check focuses on the informational content of each point, while the space reduction/merging propositions depending on the actual output values returned by tested points in various regions.