End of Day 5/12
Have organized and restructured a lot of the checkSpace class and I currently have methods (uncommented) that are able to use scikit learn's decision tree regressor to create a decision tree from the Input and Output data passed to it.  I think the way to go about it later will be to manipulate the output values of the decision tree being learned to involved failure amounts and such.  But currently the code works and the script gathers all of the inequalities of the decision tree in the irules_new list