End of Day 5/24
Working code as of the end of the day on 5/24/23.  But definitely need to go back through commenting, reorganizing, old code and so forth and of course testing it all.  But the sortPoints function is complete moving failing points to an eliminated key of each disciplines dictionary and the new rules being added have had their signs flipped and organized in an Or relational.