Script updates
Updates primarily to the plotting at the end of the script file where there was a bug in what axes were being plotted when in the last discipline.  Graphs are accurate now and line up much better with the space remaining for that last discipline.  Prior to disciplines were fine because I defaults to solving for the x3 and x5 in those disciplines.  While in the third discipline, I am solving for x1 and so need to plot j, k, and l accordingly.