Risk Quantified in Code
The code is able to quantify risk, but it is often calculating risk from a reduced potential rather than an added risk.  Need to perhaps rethink the logic for this or how I would want to present it.  Code is working as is though right now and results are certainly interesting.